Move Piece into its own module
